requirements.txt
siwe
web3
behave==1.2.6
behavex==1.5.11
faker
pyyaml==5.3
fauxfactory==3.0.6
parse==1.19.0
reportportal-client==5.0.6
requests==2.24.0
jsonschema==3.2.0
jsondiff==2.0.0
MaildropccReader
